requirements.txt
psutil
tk
pillow